backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Web application backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic[email]>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "python-jose[cryptography]>=3.3,<4",
    "bcrypt>=4.0,<5",
    "python-multipart>=0.0.18",
    "httpx>=0.28,<1",
    "psycopg2-binary>=2.9,<3",
    "slowapi>=0.1.9,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "black>=24.10,<25",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "T20", # flake8-print
    "SIM", # flake8-simplify
]
ignore = ["S105", "S106"]  # Ignore hardcoded password warnings in config

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
target-version = ["py312"]
line-length = 88

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
